obs-lab/pe-lab/streamlit-app/requirements.txt
streamlit>=1.35
kafka-python
pandas